chore: Update IP to use storage
