--- ice-phishing/src/keys.js
@@ -1,16 +1,0 @@
-const keys = {
-  etherscanApiKeys: [],
-  optimisticEtherscanApiKeys: [],
-  bscscanApiKeys: [],
-  polygonscanApiKeys: [],
-  fantomscanApiKeys: [],
-  arbiscanApiKeys: [],
-  snowtraceApiKeys: [],
-};
-
-const ZETTABLOCK_API_KEY = "";
-
-module.exports = {
-  keys,
-  ZETTABLOCK_API_KEY
-};
